LayoutTests/fast/web-share/share-transient-activation-expired-expected.txt

PASS: Share promise was rejected: NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission.
